// carbackend/src/permissions/permission.enum.ts
/* eslint-disable prettier/prettier */
// src/permissions/permission.enum.ts
export enum Permission {
       MANAGE_VEHICLES=  'MANAGE_VEHICLES',
        VIEW_VEHICLES = 'VIEW_VEHICLES',
        BOOK_VEHICLE = ' BOOK_VEHICLE',
        VIEW_BOOKINGS = 'VIEW_BOOKINGS',
        MANAGE_BOOKINGS = 'MANAGE_BOOKINGS',
        MANAGE_USERS = 'MANAGE_USERS',
        VIEW_USERS = 'VIEW_USERS',
        APPROVE_BOOKINGS = 'APPROVE_BOOKINGS',
        CANCEL_BOOKING = 'CANCEL_BOOKING',
        UPDATE_BOOKING_STATUS = 'UPDATE_BOOKING_STATUS',
        
      }
      
  